### 2019-05-24 #307 @eebrown
Submission: chlorpromazineR
Submitting Author: Eric Brown (@eebrown)
Repository: https://github.com/eebrown/chlorpromazineR
Version submitted: 0.1.0
Version accepted: TBD

---

```
Package: chlorpromazineR
Title: Convert Antipsychotic Doses to Chlorpromazine Equivalents
Version: 0.1.0
    c(person(given = "Eric",
             family = "Brown",
             email = "eb@ericebrown.com"),
      person(given = "Parita",
             family = "Shah",
             role = "aut"),
      person(given = "Julia",
             family = "Kim",
             role = "aut"))
Description: As different antipsychotic medications have different potencies,
    the doses of different medications cannot be directly compared. Various
    strategies are used to convert doses into a common reference so that
    comparison is meaningful. Chlorpromazine (CPZ) has historically been used
    as a reference medication into which other antipsychotic doses can be
    converted, as "chlorpromazine-equivalent doses". Using conversion keys
    generated from widely-cited scientific papers (Gardner et. al 2010
    <https://doi.org/10.1176/appi.ajp.2009.09060802>, Leucht et al. 2016
    <https://doi.org/10.1093/schbul/sbv167>), antipsychotic doses are converted
    to CPZ (or any specified antipsychotic) equivalents. The use of the package
    is described in the included vignette. Not for clinical use.
License: GPL-3 + file LICENSE
    covr
```

(See #306 - accepted presubmission inquiry) This package falls within "data munging" by virtue of its main feature - to convert antipsychotic doses into a common reference equivalent. It fosters reproducibility as the current practice is typically to reference the source paper without showing the conversion method (which if done manually, though simple, is prone to error). This improves accuracy and transparency.

Researchers in psychiatry/neuroscience involving antipsychotic medications often need to calculate equivalent doses, which is why the papers describing such methods are widely cited. Yet, there is no R package that facilitates this.

None to our knowledge.

#306 @noamross
