### 2019-05-24 #307 @karthik
### Editor checks:

---

#### Editor comments

👋 @eebrown! I'll be your editor on this submission. A goodpractice report on your packages comes out mostly clean! Once I assign reviewers, we'll take a deeper dive through your package. Here are some small housekeeping things to do in the meantime.

- Add a badge to your repo. It will update as the review progresses and change to 'reviewed' once accepted.\

I'd encourage you to register for an  [ORCID](https://orcid.org/register) and add it to your description for each author. See more [here](https://ropensci.github.io/dev_guide/building.html)

- There is a small typo in the README.
> With this package, he hope:

he should be we?

I'll start looking for reviewers now. If you have suggestions on people who might be qualified to this (but have no conflict of interest), feel free to suggest names.
