### 2019-05-24 #307 @eebrown
Thanks, @karthik! I've updated the DESCRIPTION and README files.
