### 2019-08-09 #307 @karthik
Assigning @chasemc as reviewer 1. Please see guidelines here: https://devguide.ropensci.org/reviewerguide.html
