### 2019-08-15 #307 @karthik
Assigning @fboehm as reviewer 2
Please see guidelines here: devguide.ropensci.org/reviewerguide.html
