### 2019-08-28 #307 @chasemc
*Chase's note- My review and inserted links correspond to the chlorpromazineR repository at git commit: `f8a0f152fc57ec0afb0c60910a9764dee780508f`*

  - https://github.com/eebrown/chlorpromazineR/blame/master/README.md#L5-L11
  - https://github.com/eebrown/chlorpromazineR/blame/master/README.md#L19
  - https://github.com/eebrown/chlorpromazineR/blob/master/vignettes/walkthrough.Rmd
  - https://github.com/eebrown/chlorpromazineR/tree/master/man
  - https://travis-ci.org/eebrown/chlorpromazineR/builds/536968685#L1026

##### For packages co-submitting to JOSS

- NA

Session info for installation test:

```
> sessionInfo()
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 17763)

[1] LC_COLLATE=English_United States.1252
[2] LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252
[4] LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

 [1] Rcpp_1.0.2        ps_1.3.0          prettyunits_1.0.2 rprojroot_1.3-2
 [5] digest_0.6.20     crayon_1.3.4      withr_2.1.2       assertthat_0.2.1
 [9] R6_2.4.0          backports_1.1.4   magrittr_1.5      rlang_0.4.0
[13] cli_1.1.0         curl_4.0          fs_1.3.1          remotes_2.1.0
[17] testthat_2.2.1    callr_3.3.1       desc_1.2.0        devtools_2.1.0
[21] tools_3.6.1       glue_1.3.1        pkgload_1.0.2     compiler_3.6.1
[25] processx_3.4.1    pkgbuild_1.0.4    sessioninfo_1.1.1 memoise_1.1.0
[29] usethis_1.5.1
```

  - Vignette and examples were reproduced. Also, the unit tests are sufficient to ensure the authors are recieving expected results (e.g. [hard-coded](https://github.com/eebrown/chlorpromazineR/blob/f8a0f152fc57ec0afb0c60910a9764dee780508f/tests/testthat/test_convert.R#L7-L22) expected results are tested against).

   and a reasonable range of inputs and conditions. All tests pass on my local machine.

```
==> devtools::test()

Loading chlorpromazineR
Testing chlorpromazineR
v |  OK F W S | Context
v |  19       | checking functions behave [0.1 s]
v |  10       | CPZ-equivalent conversion
v |  15       | checking key manipulation functions

== Results =============================

OK:       44
```

Estimated hours spent reviewing: 5 (partly getting used to/reading rOpenSci procedure)

---

**Community guidelines**

 - Contribution guidelines are missing.

**Packaging guidelines**

 - Not sure if this is in the guidelines but I would recommend removing the license section from the README as it is not self-sufficient and may confuse potential users.

 - "Brief demonstration usage" should be added to the README
   - https://devguide.ropensci.org/building.html#readme
   - Can probably just copy the example from chlorpromazineR::to_cpz()

I would suggest removing "age" from the vignettes/examples, as it suggests it might be used in the conversion. Alternately, change to another "metadata" that obviously wouldn't be used in dosing decisions, or make it clear that age isn't accounted for in the conversion.

**Error Messages**

Consider making a clear error message for likely things like someone providing a route that isn't accepted:
```
> to_cpz(example_oral, ap_label = "antipsychotic", dose_label = "dose",
+        route = "nope")

Error in check_ap(x, key = key, ap_label = ap_label, route = route, route_label = route_label) :
  object 'notfound' not found
```

**Code Style**
 There are some places that could benefit from writing code in clearer style, such as using spaces- like [here](https://github.com/eebrown/chlorpromazineR/blob/f8a0f152fc57ec0afb0c60910a9764dee780508f/R/convert.R#L100-L104) (not required but rOpenSci does recommend http://r-pkgs.had.co.nz/r.html). And, while not incorrect, it would make the code easier trace if [that same section](https://github.com/eebrown/chlorpromazineR/blob/f8a0f152fc57ec0afb0c60910a9764dee780508f/R/convert.R#L100-L104) was broken into steps and/or code comments were added.

**Previous Similar Work**
There has been some previous similar work, the Stefan Leucht paper cited in this package (https://dx.doi.org/10.1093%2Fschbul%2Fsbv167) links to an excel tool "Antipsychotic dose conversion website.xls" available from:
http://www.cfdm.de/indexab2e.html?option=com_content&task=view&id=15&Itemid=29
While the Leucht paper is cited as a source of data, the tool is not and probably should be.

**Overall**
Otherwise, the package was well laid out and had low dependencies and therefore limited worry about breaking future changes.
