### 2019-09-02 #307 @fboehm
Overall, I like this package and think that it is useful and well written. I make a few suggestions below that, I hope, may improve the user interface and ease of use. I also include the R code (and output) for the checks that I ran.

Being a new to the package, I tried to think about ways that a new user might guess incorrectly about function use. I encourage the authors to clarify error messages so as to guide the new user towards the correct way to specify the functions' inputs.

While the authors state that the package is not for clinical use, I didn't notice an explicit statement of who the target audience is.

I don't know who the maintainer is. Maybe this information can be made explicit in the DESCRIPTION file?

Also, I don't see clear guidelines for contributions.

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

## **Reviewer:** [\@fboehm](https://github.com/fboehm)

### Review Submitted:
****

***

This report contains documents associated with the review of **rOpenSci** submitted package:

## Package info

**Description:**

As different antipsychotic medications have different potencies,
    the doses of different medications cannot be directly compared. Various
    strategies are used to convert doses into a common reference so that
    comparison is meaningful. Chlorpromazine (CPZ) has historically been used
    as a reference medication into which other antipsychotic doses can be
    converted, as "chlorpromazine-equivalent doses". Using conversion keys
    generated from widely-cited scientific papers (Gardner et. al 2010
    <https://doi.org/10.1176/appi.ajp.2009.09060802>, Leucht et al. 2016
    <https://doi.org/10.1093/schbul/sbv167>), antipsychotic doses are converted
    to CPZ (or any specified antipsychotic) equivalents. The use of the package
    is described in the included vignette. Not for clinical use.

**Author:** Eric Brown <eb@ericebrown.com> [aut, cre] (<https://orcid.org/0000-0002-1575-2606>), Parita Shah [aut] (<https://orcid.org/0000-0002-7302-0411>), Julia Kim [aut] (<https://orcid.org/0000-0002-0379-1333>)

**repo url:** <https://github.com/eebrown/chlorpromazineR>

**website url:** <>

## Review info

#### See [reviewer guidelines](https://ropensci.github.io/dev_guide/reviewerguide.html) for further information on the rOpenSci review process.

**key review checks:**

- Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?

The Mozilla guide divides review questions into two categories: Intrinsic and Extrinsic. I reproduce this structure below.

*Intrinsic examination*

1. Are functions as simple as possible?

Yes, the functions are as simple as possible.

1. Is the code efficient?

Yes.

1. Is the usage of each function clear?

Yes, the functions have clear usages.

1. Have edge cases been considered?

I'm not sure what counts as an edge case in this setting. The checks that run on inputs to `to_cpz`, and possibly other functions, may be insufficient. While I recognize that the `to_cpz` help file tells the user that argument `x` is a dateframe, I think it would be helpful for `to_cpz` code to verify that `x` is a dateframe. For example, the base function `is.data.frame` might be used.

*Extrinsic examination*

1. Does the new code reinvent any wheels?

No.

1. Does the new code successfully address the needs of the project?

Yes, I think it does.

1. Does the new code respect the structure of the project?

Yes.

I'm not sure where to comment on the data. My question is "Would it be useful to provide the data as csv files in addition to rda files?" It looks like you have a directory, "extdata", that contains csv files, yet "extdata" is not part of the submitted package. My thought is that a user may wish to verify the translation and data entry of the tables from the publications into the csv files. (I'm assuming that you entered the results by hand by reading the values from the publications and typing them into a csv.) Providing the csv files may make that easier to verify.

- Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?

Yes, it complies.

- Are there **improvements** that could be made to the **code style?**

No. In reading the code in the three files within the R/ directory, I feel that the code style is satisfactory.

- Is there **code duplication** in the package that should be reduced?

No, I didn't see any code duplication.

- Are there **user interface improvements** that could be made?

Would it be sensible to add a column of units to dataframes in your examples in the vignette? You may want to add a column for dosage units and a column for units on the "q" variable. Also, is there a better (more informative) name for the "q" variable? Maybe "frequency" or "period"?

I also suggest adding example code for the vignette section "Participants on multiple medications". I'm unsure what the guidelines suggest for those on multiple anti-psychotics. When computing "chlorpromazine-equivalent" for a participant on multiple drugs, does one merely add the chlorpromazine-equivalent values for each drug?

Also, regarding the names of drugs, is it reasonable to assume that a given drug is only available as a single "salt"? For example, when speaking of "fluphenazine", you use the term "fluphenazine hcl". What if a subject takes fluphenazine in a compound that is not "fluphenazine hcl"? It seems that one would need distinct conversion factors for each compound. For this reason, I am unsure whether I like the use of functions that "trim" the drug names.

- Are there **performance improvements** that could be made?

No. The calculations are straightforward. I don't see how performance could be improved.

- Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?

Documentation is both clear and sufficient.

Please be respectful and kind to the authors in your reviews. The rOpenSci [code of conduct](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) is mandatory for everyone involved in our review process.

***

### session info

```r
#> R version 3.6.1 (2019-07-05)
#> Platform: x86_64-apple-darwin15.6.0 (64-bit)
#> Running under: macOS Mojave 10.14.6
#>
#> Matrix products: default
#> BLAS:   /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRblas.0.dylib
#> LAPACK: /Library/Frameworks/R.framework/Versions/3.6/Resources/lib/libRlapack.dylib
#>
#> locale:
#> [1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8
#>
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base
#>
#> other attached packages:
#> [1] devtools_2.1.0 usethis_1.5.1  magrittr_1.5
#>
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_1.0.2.1        knitr_1.24          pkgload_1.0.2
#>  [4] R6_2.4.0            rlang_0.4.0         stringr_1.4.0
#>  [7] tools_3.6.1         pkgbuild_1.0.5.9000 xfun_0.9
#> [10] sessioninfo_1.1.1   cli_1.1.0           withr_2.1.2
#> [13] remotes_2.1.0       htmltools_0.3.6     rprojroot_1.3-2
#> [16] yaml_2.2.0          digest_0.6.20       assertthat_0.2.1
#> [19] crayon_1.3.4        processx_3.4.1      callr_3.3.1
#> [22] fs_1.3.1            ps_1.3.0            testthat_2.2.1
#> [25] memoise_1.1.0       glue_1.3.1          evaluate_0.14
#> [28] rmarkdown_1.15      stringi_1.4.3       compiler_3.6.1
#> [31] backports_1.1.4     desc_1.2.0          prettyunits_1.0.2
```

## Test installation

### test local `chlorpromazineR` install:

```r
install(pkg_dir, dependencies = TRUE, build_vignettes = TRUE)
#>
#>
   checking for file ‘/Users/frederickboehm/work-2019/peer-review/ropensci/reviews/chlorpromazineR/DESCRIPTION’ ...

✔  checking for file ‘/Users/frederickboehm/work-2019/peer-review/ropensci/reviews/chlorpromazineR/DESCRIPTION’
#>

─  preparing ‘chlorpromazineR’:
#>

   checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#>

#>

   creating vignettes ...

✔  creating vignettes (2.5s)
#>

#>

#>

─  looking to see if a ‘data/datalist’ file should be added
#>

─  building ‘chlorpromazineR_0.1.0.tar.gz’
#>

#>
Running /Library/Frameworks/R.framework/Resources/bin/R CMD INSTALL \
#>   /var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T//RtmpboPMAb/chlorpromazineR_0.1.0.tar.gz \
#>   --install-tests
#> * installing to library ‘/Library/Frameworks/R.framework/Versions/3.6/Resources/library’
#> * installing *source* package ‘chlorpromazineR’ ...
#> ** using staged installation
#> ** R
#> ** data
#> *** moving datasets to lazyload DB
#> ** inst
#> ** tests
#> ** byte-compile and prepare package for lazy loading
#> ** help
#> *** installing help indices
#> ** building package indices
#> ** installing vignettes
#> ** testing if installed package can be loaded from temporary location
#> ** testing if installed package can be loaded from final location
#> ** testing if installed package keeps a record of temporary installation path
#> * DONE (chlorpromazineR)
```

```r
remove.packages("chlorpromazineR")
#> Removing package from '/Library/Frameworks/R.framework/Versions/3.6/Resources/library'
#> (as 'lib' is unspecified)
```
#### **comments:**

Easily installed with no warnings or errors. No R package dependencies.

***

### test install of `chlorpromazineR` from GitHub with:

```r
devtools::install_github("eebrown/chlorpromazineR", dependencies = TRUE, build_vignettes = TRUE)
#> Downloading GitHub repo eebrown/chlorpromazineR@master
#>
#>
   checking for file ‘/private/var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T/RtmpboPMAb/remoteseff855670471/eebrown-chlorpromazineR-f8a0f15/DESCRIPTION’ ...

✔  checking for file ‘/private/var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T/RtmpboPMAb/remoteseff855670471/eebrown-chlorpromazineR-f8a0f15/DESCRIPTION’
#>

─  preparing ‘chlorpromazineR’:
#>

   checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#>

#>

   creating vignettes ...

✔  creating vignettes (2.2s)
#>

#>

#> ─  looking to see if a ‘data/datalist’ file should be added
#> ─  building ‘chlorpromazineR_0.1.0.tar.gz’
#>

#>
```

#### **comments:**

Easily installed.

***

## Check package integrity

### run checks on `chlorpromazineR` source:

```r
devtools::check(pkg_dir)
#> Updating chlorpromazineR documentation
#> Writing NAMESPACE
#> Loading chlorpromazineR
#> Writing NAMESPACE
#> ── Building ─────────────────────────────────────────────────────────────────────── chlorpromazineR ──
#> Setting env vars:
#> ● CFLAGS    : -Wall -pedantic
#> ● CXXFLAGS  : -Wall -pedantic
#> ● CXX11FLAGS: -Wall -pedantic
#> ──────────────────────────────────────────────────────────────────────────────────────────────────────
#>
   checking for file ‘/Users/frederickboehm/work-2019/peer-review/ropensci/reviews/chlorpromazineR/DESCRIPTION’ ...

✔  checking for file ‘/Users/frederickboehm/work-2019/peer-review/ropensci/reviews/chlorpromazineR/DESCRIPTION’
#>

─  preparing ‘chlorpromazineR’:
#>

   checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#>

#>

   creating vignettes ...

✔  creating vignettes (2.4s)
#>

#>

#> ─  looking to see if a ‘data/datalist’ file should be added
#>

─  building ‘chlorpromazineR_0.1.0.tar.gz’
#>

#>
── Checking ─────────────────────────────────────────────────────────────────────── chlorpromazineR ──
#> Setting env vars:
#> ● _R_CHECK_CRAN_INCOMING_REMOTE_: FALSE
#> ● _R_CHECK_CRAN_INCOMING_       : FALSE
#> ● _R_CHECK_FORCE_SUGGESTS_      : FALSE
#> ── R CMD check ────────────────────────────────────────────────────────────
#> * using log directory ‘/private/var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T/RtmpboPMAb/chlorpromazineR.Rcheck’
#> * using R version 3.6.1 (2019-07-05)
#> * using platform: x86_64-apple-darwin15.6.0 (64-bit)
#> * using session charset: UTF-8
#> * using options ‘--no-manual --as-cran’
#> * checking for file ‘chlorpromazineR/DESCRIPTION’ ... OK
#> * this is package ‘chlorpromazineR’ version ‘0.1.0’
#> * package encoding: UTF-8
#> * checking package namespace information ... OK
#> * checking package dependencies ... OK
#> * checking if this is a source package ... OK
#> * checking if there is a namespace ... OK
#> * checking for executable files ... OK
#> * checking for hidden files and directories ... OK
#> * checking for portable file names ... OK
#> * checking for sufficient/correct file permissions ... OK
#> * checking whether package ‘chlorpromazineR’ can be installed ... OK
#> * checking installed package size ... OK
#> * checking package directory ... OK
#> * checking for future file timestamps ... OK
#> * checking ‘build’ directory ... OK
#> * checking DESCRIPTION meta-information ... OK
#> * checking top-level files ... OK
#> * checking for left-over files ... OK
#> * checking index information ... OK
#> * checking package subdirectories ... OK
#> * checking R files for non-ASCII characters ... OK
#> * checking R files for syntax errors ... OK
#> * checking whether the package can be loaded ... OK
#> * checking whether the package can be loaded with stated dependencies ... OK
#> * checking whether the package can be unloaded cleanly ... OK
#> * checking whether the namespace can be loaded with stated dependencies ... OK
#> * checking whether the namespace can be unloaded cleanly ... OK
#> * checking dependencies in R code ... OK
#> * checking S3 generic/method consistency ... OK
#> * checking replacement functions ... OK
#> * checking foreign function calls ... OK
#> * checking R code for possible problems ... OK
#> * checking Rd files ... OK
#> * checking Rd metadata ... OK
#> * checking Rd line widths ... OK
#> * checking Rd cross-references ... OK
#> * checking for missing documentation entries ... OK
#> * checking for code/documentation mismatches ... OK
#> * checking Rd \usage sections ... OK
#> * checking Rd contents ... OK
#> * checking for unstated dependencies in examples ... OK
#> * checking contents of ‘data’ directory ... OK
#> * checking data for non-ASCII characters ... OK
#> * checking data for ASCII and uncompressed saves ... OK
#> * checking installed files from ‘inst/doc’ ... OK
#> * checking files in ‘vignettes’ ... OK
#> * checking examples ... OK
#> * checking for unstated dependencies in ‘tests’ ... OK
#> * checking tests ...
#>   Running ‘testthat.R’
#>  OK
#> * checking for unstated dependencies in vignettes ... OK
#> * checking package vignettes in ‘inst/doc’ ... OK
#> * checking re-building of vignette outputs ... OK
#> * DONE
#>
#> Status: OK
#> ── R CMD check results ───────────────────────── chlorpromazineR 0.1.0 ────
#> Duration: 16.7s
#>
#> 0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```
#### **comments:**

***

### run tests on `chlorpromazineR` source:

```r
devtools::test(pkg_dir)
#> Loading chlorpromazineR
#>
#> Attaching package: 'testthat'
#> The following object is masked from 'package:devtools':
#>
#>     test_file
#> The following objects are masked from 'package:magrittr':
#>
#>     equals, is_less_than, not
#> Testing chlorpromazineR
#> ✔ |  OK F W S | Context
#>
⠏ |   0       | checking functions behave
✔ |  19       | checking functions behave
#>
⠏ |   0       | CPZ-equivalent conversion
✔ |  10       | CPZ-equivalent conversion
#>
⠏ |   0       | checking key manipulation functions
✔ |  15       | checking key manipulation functions
#>
#> ══ Results ═══════════════════════════════════════════════════════════════════════════════════════════
#> Duration: 0.2 s
#>
#> OK:       44
#> Failed:   0
#> Warnings: 0
#> Skipped:  0
```
#### **comments:**

The tests all pass.

***

### check `chlorpromazineR` for goodpractice:

```r
goodpractice::gp(pkg_dir)
#> Preparing: covr
#> Preparing: cyclocomp
#>
   checking for file ‘/private/var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T/RtmpboPMAb/remoteseff81a5e5b30/chlorpromazineR/DESCRIPTION’ ...

✔  checking for file ‘/private/var/folders/wd/lxmyvz590xb81c5z1j88b3800000gn/T/RtmpboPMAb/remoteseff81a5e5b30/chlorpromazineR/DESCRIPTION’
#>

─  preparing ‘chlorpromazineR’:
#>

   checking DESCRIPTION meta-information ...

✔  checking DESCRIPTION meta-information
#>

   checking vignette meta-information ...

✔  checking vignette meta-information
#>

#>

#>

─  looking to see if a ‘data/datalist’ file should be added
#>

─  building ‘chlorpromazineR_0.1.0.tar.gz’
#>

#>
#> Preparing: description
#> Preparing: lintr
#> Preparing: namespace
#> Preparing: rcmdcheck
#> ── GP chlorpromazineR ─────────────────────────────────────────────────────
#>
#> It is good practice to
#>
#>   ✖ write unit tests for all functions, and all package code in
#>     general. 90% of code lines are covered by test cases.
#>
#>     R/convert.R:48:NA
#>     R/convert.R:60:NA
#>     R/convert.R:67:NA
#>     R/convert.R:72:NA
#>     R/convert.R:73:NA
#>     ... and 6 more lines
#>
#> ───────────────────────────────────────────────────────────────────────────
```
#### **comments:**

***

## Check package metadata files

### inspect

- #### [README](https://github.com/eebrown/chlorpromazineR)
- #### [DESCRIPTION](https://github.com/eebrown/chlorpromazineR/blob/master/DESCRIPTION)
- #### [NAMESPACE](https://github.com/eebrown/chlorpromazineR/blob/master/NAMESPACE)

### spell check

```r
devtools::spell_check(pkg_dir)
#> DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
```

  <script data-pagedtable-source type="application/json">
[…]
