### 2019-09-06 #307 @eebrown
Dear @chasemc and @fboehm, thank you both very much for your thorough and helpful reviews. I will address your suggestions in an update to the package. Unfortunately the timing has not worked out well and I will be out of office for approx. 2 weeks, but plan to update the package quickly upon my return. Warm regards, Eric
