### 2019-09-25 #307 @eebrown
Dear @chasemc - thanks again for your helpful and thorough review. I've addressed your comments below:

> ### Review Comments
> **Community guidelines**
>
> * Contribution guidelines are missing.

Thanks - I've added contributing guidelines and templates with rOpenSci’s use_ro_github() - https://github.com/eebrown/chlorpromazineR/commit/db1fa2b090140a150c04f325fd83f348741488c4

> **Packaging guidelines**
>
> * Not sure if this is in the guidelines but I would recommend removing the license section from the README as it is not self-sufficient and may confuse potential users.

Good point. I've removed license text from README.md. https://github.com/eebrown/chlorpromazineR/commit/8d2f9c52bbf3c6d5a31e48923db81130938a121c

> * "Brief demonstration usage" should be added to the README
>
>   * https://devguide.ropensci.org/building.html#readme
>   * Can probably just copy the example from chlorpromazineR::to_cpz()

Agreed and added.  https://github.com/eebrown/chlorpromazineR/commit/307e23f40311de003c06fc4ae17d3d7755b3718e

> I would suggest removing "age" from the vignettes/examples, as it suggests it might be used in the conversion. Alternately, change to another "metadata" that obviously wouldn't be used in dosing decisions, or make it clear that age isn't accounted for in the conversion.

I agree that it could be confusing, so I added a clarifying comment. The reason I included it is to show that the data.frame can include all sorts of extra variables, and age is a typical one. https://github.com/eebrown/chlorpromazineR/commit/8dec24ca7eed7cdc6f3d53cac62178a7c6b8a065

> **Error Messages**
>
> Consider making a clear error message for likely things like someone providing a route that isn't accepted:
>
> ```
> > to_cpz(example_oral, ap_label = "antipsychotic", dose_label = "dose",
> +        route = "nope")
>
> Error in check_ap(x, key = key, ap_label = ap_label, route = route, route_label = route_label) :
>   object 'notfound' not found
> ```

Thanks, great point. I've added a help function `check_params()` to check that the parameters are the expected format. https://github.com/eebrown/chlorpromazineR/commit/da2ced87d452787b76c608147c30ca444146181c https://github.com/eebrown/chlorpromazineR/commit/64b9b89f21bc936096d08a64195cfbe99cb43c52

> **Code Style**
> There are some places that could benefit from writing code in clearer style, such as using spaces- like [here](https://github.com/eebrown/chlorpromazineR/blob/f8a0f152fc57ec0afb0c60910a9764dee780508f/R/convert.R#L100-L104) (not required but rOpenSci does recommend http://r-pkgs.had.co.nz/r.html). And, while not incorrect, it would make the code easier trace if [that same section](https://github.com/eebrown/chlorpromazineR/blob/f8a0f152fc57ec0afb0c60910a9764dee780508f/R/convert.R#L100-L104) was broken into steps and/or code comments were added.

I've added spacing per the recommendation. https://github.com/eebrown/chlorpromazineR/commit/b241b9f55e7fd2d1879bef2a03d7cfed95eb443c

> **Previous Similar Work**
> There has been some previous similar work, the Stefan Leucht paper cited in this package (https://dx.doi.org/10.1093%2Fschbul%2Fsbv167) links to an excel tool "Antipsychotic dose conversion website.xls" available from:
> http://www.cfdm.de/indexab2e.html?option=com_content&task=view&id=15&Itemid=29
> While the Leucht paper is cited as a source of data, the tool is not and probably should be.

Good idea, thanks. Citations added to README. https://github.com/eebrown/chlorpromazineR/commit/ab21898e16e222cc22c11c7969737471b0b9b175

> **Overall**
> Otherwise, the package was well laid out and had low dependencies and therefore limited worry about future breaking changes.

Thanks again. Please let me know if you feel I did not fully understand or address any of your feedback and I am happy to revisit.

Responses to @fboehm coming soon...

Best,
@eebrown
