### 2019-09-25 #307 @eebrown
Dear @fboehm,

Thanks very much for your review. Please find my responses below:

> Overall, I like this package and think that it is useful and well written. I make a few suggestions below that, I hope, may improve the user interface and ease of use. I also include the R code (and output) for the checks that I ran.

Thanks 👍

> Being a new to the package, I tried to think about ways that a new user might guess incorrectly about function use. I encourage the authors to clarify error messages so as to guide the new user towards the correct way to specify the functions' inputs.

I believe this has been addressed with the new `check_params()` function. https://github.com/eebrown/chlorpromazineR/commit/da2ced87d452787b76c608147c30ca444146181c https://github.com/eebrown/chlorpromazineR/commit/64b9b89f21bc936096d08a64195cfbe99cb43c52

> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
> While the authors state that the package is not for clinical use, I didn't notice an explicit statement of who the target audience is.

I've added a statement to the README. The target audience is scientists doing clinical research involving antipsychotic medications. https://github.com/eebrown/chlorpromazineR/commit/9a167848b5a3a66e420f3a7842824e8932ff6725

> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
> * [x]  **Function Documentation:** for all exported functions in R help
> * [x]  **Examples** for all exported functions in R Help that run successfully locally
> * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> I don't know who the maintainer is. Maybe this information can be made explicit in the DESCRIPTION file?

This is already in the description file. The [cre] role indicates maintainer, per http://r-pkgs.had.co.nz/description.html.

> Also, I don't see clear guidelines for contributions.

Thanks - I have now added these using the rOpenSci function use_ro_github() https://github.com/eebrown/chlorpromazineR/commit/db1fa2b090140a150c04f325fd83f348741488c4

> ...

> ## Review info
> #### See [reviewer guidelines](https://ropensci.github.io/dev_guide/reviewerguide.html) for further information on the rOpenSci review process.
> **key review checks:**
>
> * Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
>
> The Mozilla guide divides review questions into two categories: Intrinsic and Extrinsic. I reproduce this structure below.
>
> _Intrinsic examination_
>
> 1. Are functions as simple as possible?
>
> Yes, the functions are as simple as possible.
>
> 1. Is the code efficient?
>
> Yes.
>
> 1. Is the usage of each function clear?
>
> Yes, the functions have clear usages.
>
> 1. Have edge cases been considered?
>
> I'm not sure what counts as an edge case in this setting. The checks that run on inputs to `to_cpz`, and possibly other functions, may be insufficient. While I recognize that the `to_cpz` help file tells the user that argument `x` is a dateframe, I think it would be helpful for `to_cpz` code to verify that `x` is a dateframe. For example, the base function `is.data.frame` might be used.

As noted above, this has been addressed with `check_params()` helper function.

> _Extrinsic examination_
>
> 1. Does the new code reinvent any wheels?
>
> No.
>
> 1. Does the new code successfully address the needs of the project?
>
> Yes, I think it does.
>
> 1. Does the new code respect the structure of the project?
>
> Yes.
>
> I'm not sure where to comment on the data. My question is "Would it be useful to provide the data as csv files in addition to rda files?" It looks like you have a directory, "extdata", that contains csv files, yet "extdata" is not part of the submitted package. My thought is that a user may wish to verify the translation and data entry of the tables from the publications into the csv files. (I'm assuming that you entered the results by hand by reading the values from the publications and typing them into a csv.) Providing the csv files may make that easier to verify.

I agree and encourage users to verify the data entry. I included the external data in inst/extdata as recommended in http://r-pkgs.had.co.nz/data.html. I thought this made the most sense, as it is stored in the repository for anyone to verify.

> * Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
>
> Yes, it complies.
>
> * Are there **improvements** that could be made to the **code style?**
>
> No. In reading the code in the three files within the R/ directory, I feel that the code style is satisfactory.
>
> * Is there **code duplication** in the package that should be reduced?
>
> No, I didn't see any code duplication.
>
> * Are there **user interface improvements** that could be made?
>
> Would it be sensible to add a column of units to dataframes in your examples in the vignette? You may want to add a column for dosage units and a column for units on the "q" variable. Also, is there a better (more informative) name for the "q" variable? Maybe "frequency" or "period"?

I chose not to add units, because antipsychotics are universally prescribed in mg. Also, as long as all units are the same, the conversion factors would still work.

I chose the name “q”, because it is actually descriptive, e.g. the LAIs are dosed q 2 weeks, q 3 weeks, etc. It could equally be frequency, but q is nice a short.

> I also suggest adding example code for the vignette section "Participants on multiple medications". I'm unsure what the guidelines suggest for those on multiple anti-psychotics. When computing "chlorpromazine-equivalent" for a participant on multiple drugs, does one merely add the chlorpromazine-equivalent values for each drug?

As far as I know there are not guidelines and the user should decide how best to treat this scenario depending on their study/rationale.

> Also, regarding the names of drugs, is it reasonable to assume that a given drug is only available as a single "salt"? For example, when speaking of "fluphenazine", you use the term "fluphenazine hcl". What if a subject takes fluphenazine in a compound that is not "fluphenazine hcl"? It seems that one would need distinct conversion factors for each compound. For this reason, I am unsure whether I like the use of functions that "trim" the drug names.

It is almost always the case that there is only one compound, so this function is provided for convenience when the names in the user’s data only use the main name. But as you point out, it is a major issue if it introduces ambiguity between two compounds. The user must verify that this is not the case before using the trim function. I’ve added a message to the trim_function to warn the user.  https://github.com/eebrown/chlorpromazineR/commit/170fc76cac9b246798ec84add653541def756618
> * Are there **performance improvements** that could be made?
>
> No. The calculations are straightforward. I don't see how performance could be improved.
>
> * Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?
>
> Documentation is both clear and sufficient.
> ...
> #### **comments:**
> It would be useful to give a more informative message to the user when they enter something like the code above. Perhaps amending the code for `to_cpz` so that it checks whether inputs have appropriate class & dimension and/or are not `NULL` would be useful. You could write a line that checks whether `x` is a dataframe, for example.

As noted above, this has been addressed with `check_params()` helper function.

> ## Inspect code:
> * Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
>
>   * good [function & variable naming?](https://ropensci.github.io/dev_guide/building.html#function-and-argument-naming)
>
>   The function argument names are well chosen for the most part. Would it be better to rename `x` in the function `to_cpz`? Maybe calling it `dataframe` would work better.

I think in rOpenSci and tidyverse functions it is fairly standard to use `x` as the first parameters/object of a function as the data object that is being acted on. In this way it is clear what is being changed, e.g. when using the pipe operator.

> My other (minor) issue is the use of "q" as a variable name in the vignette.

I chose the name “q”, because it is actually descriptive, e.g. the LAIs are dosed q 2 weeks, q 3 weeks, etc. It could equally be frequency, but q is nice a short.

Thanks again and please also let me know if you feel I did not fully understand or address any of your feedback and I am happy to revisit.

Best,
@eebrown
