### 2019-09-26 #307 @fboehm
Thanks, @eebrown. I was mistaken about the creator/maintainer being unlisted. Thanks for the correction. I'm still concerned about the variable naming conventions that you use. Maybe it's just my preference that the variables be informatively named, but I especially dislike the use of "x" as a variable name. As a new user, "x" doesn't tell me what the input needs to be. If you prefer to maintain the use of "q" as a variable, that's fine, although I prefer "period" (the inverse of frequency).
