### 2019-09-26 #307 @chasemc
Looks good on my comments @eebrown, I've accepted your corrections and recommend approving.

@fboehm, I tend to agree on your comments about `x` and `q`.  However, in the case of `q` I think the target users will understand, as it's a common term when dealing with medications. That being said, and while not incorrect to do, I tend to avoid `q` because it is a base function: `base::q()`.
