### 2019-09-26 #307 @eebrown
Thanks @fboehm and @chasemc. I've changed the `x` argument to `input_data`. I've changed the `q` argument to `q_label` to avoid interference with `base::q()` and for symmetry with the other arguments that refer to column names in the input data.frame. I've updated the documentation, vignettes and tests accordingly as well.
