### 2019-10-01 #307 @eebrown
* [x]  Fix any links in badges for CI and coverage to point to the ropensci URL. (We'll turn on the services on our end as needed)

> Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). If you are, @stefaniebutland will be in touch about content and timing.

Thanks!
