### 2019-10-03 #307 @eebrown
Hi @stefaniebutland... it was an accident, but I do plan to prepare a tech note for chlorpromazineR, likely after I finish reviewing another package. 👍
