### 2019-10-18 #307 @karthik
I'm closing this issue but feel free to continue the conversation around crediting reviewers with their ORCID. 🙏
